+++ src/components/Collapse/Toggle.tsx
@@ -1,0 +1,23 @@
+import clsx from 'clsx';
+import { ReactNode } from 'react';
+import { Button } from '..';
+import { useCollapseContext } from './Collapse';
+
+type Props = {
+  children: ReactNode;
+  className?: string;
+};
+
+const Toggle = ({ children, className = '' }: Props): ReactNode => {
+  const { toggle } = useCollapseContext();
+
+  const classNames = clsx('toggle', className);
+
+  return (
+    <Button className={classNames} onClick={toggle}>
+      {children}
+    </Button>
+  );
+};
+
+export default Toggle;
