+++ src/components/Header/Header.tsx
@@ -1,5 +1,8 @@
 import clsx from 'clsx';
 import { FC, ReactNode } from 'react';
+import Center from './Center';
+import Left from './Left';
+import Right from './Right';
 
 type Props = {
   children: ReactNode;
@@ -11,25 +14,6 @@ interface HeaderComponent extends FC<Props> {
   Center: FC<Props>;
   Right: FC<Props>;
 }
-
-const Left = ({ children, className = '' }: Props): ReactNode => {
-  const classNames = clsx('left', className);
-
-  return <div className={classNames}>{children}</div>;
-};
-
-const Center = ({ children, className = '' }: Props): ReactNode => {
-  const styleClassNames = 'flex-1 flex justify-center';
-  const classNames = clsx('center', styleClassNames, className);
-
-  return <div className={classNames}>{children}</div>;
-};
-
-const Right = ({ children, className = '' }: Props): ReactNode => {
-  const classNames = clsx('right', className);
-
-  return <div className={classNames}>{children}</div>;
-};
 
 const Header: HeaderComponent = ({
   children,
