+++ src/components/Header/Left.tsx
@@ -1,0 +1,15 @@
+import clsx from 'clsx';
+import { ReactNode } from 'react';
+
+type Props = {
+  children: ReactNode;
+  className?: string;
+};
+
+const Left = ({ children, className = '' }: Props): ReactNode => {
+  const classNames = clsx('left', className);
+
+  return <div className={classNames}>{children}</div>;
+};
+
+export default Left;
