+++ src/components/Header/Right.tsx
@@ -1,0 +1,15 @@
+import clsx from 'clsx';
+import { ReactNode } from 'react';
+
+type Props = {
+  children: ReactNode;
+  className?: string;
+};
+
+const Right = ({ children, className = '' }: Props): ReactNode => {
+  const classNames = clsx('right', className);
+
+  return <div className={classNames}>{children}</div>;
+};
+
+export default Right;
